Extract setPath from common
